# api/docs/swagger.yaml
basePath: /
definitions:
  handlers.AskAIRequest:
    properties:
      context:
        type: string
      prompt:
        type: string
    required:
    - context
    - prompt
    type: object
  handlers.ChatRequest:
    properties:
      passage:
        minLength: 5
        type: string
      prompt:
        minLength: 2
        type: string
      themes:
        items:
          type: string
        type: array
    required:
    - passage
    - prompt
    type: object
  handlers.ChatResponse:
    properties:
      response:
        type: string
    type: object
  handlers.CreateNoteRequest:
    properties:
      content:
        additionalProperties: true
        type: object
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  handlers.Note:
    properties:
      content:
        additionalProperties: true
        type: object
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  handlers.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - full_name
    type: object
  handlers.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for the Discipleship Journal application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Discipleship Journal API
  version: "1.0"
paths:
  /api/ai/ask:
    post:
      consumes:
      - application/json
      description: Ask AI a question based on general context
      parameters:
      - description: Ask AI Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AskAIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ask AI
      tags:
      - chat
  /api/bible/passage:
    get:
      consumes:
      - application/json
      description: Fetch a bible passage from the external API
      parameters:
      - description: Bible Reference (e.g. John 3:16)
        in: query
        name: ref
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Reference is required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Bible Passage
      tags:
      - bible
  /api/chat:
    post:
      consumes:
      - application/json
      description: Chat with Bible AI using passage context
      parameters:
      - description: Chat Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Chat with AI
      tags:
      - chat
  /api/notes:
    get:
      consumes:
      - application/json
      description: Fetch all notes belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Note'
            type: array
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all notes for user
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Create a new journal note
      parameters:
      - description: Create Note Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new note
      tags:
      - notes
  /api/notes/{id}:
    get:
      consumes:
      - application/json
      description: Fetch a specific note by ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Note'
        "403":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Note not found
          schema:
            type: string
      summary: Get a single note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update an existing journal note
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Note Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Note not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a note
      tags:
      - notes
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user or update existing one based on Firebase UID
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.User'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create or update a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update existing user details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.User'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Update a user
      tags:
      - users
  /users/me:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.User'
      summary: Get current user
      tags:
      - users
swagger: "2.0"